Agent gets scared when no safe routes are available.
If the agent can't find a safe route to its destination it gets stuck. i need to find a good way to tell the agent to take chances when no safe paths are available.

Wumpus is treated as a pit for the moment.

The agent now randomly generates a destination which it then tries to go to.

Removal of potential pits which have been explored not to be pits is also added.

//TODO:
- Fix the take chances thing.
- Remove potential pits from list with help of multiple breezes.
- Treat wumpus as a wumpus and not a pit.